Reapply gateway route refactor with new changes for cosmos-to-cosmos
